+++ mats-test-junit/src/main/java/io/mats3/test/junit/Rule_MatsAnnotatedClass.java
@@ -1,11 +1,11 @@
-package io.mats3.spring.test;
+package io.mats3.test.junit;
 
-import org.junit.jupiter.api.extension.AfterEachCallback;
-import org.junit.jupiter.api.extension.BeforeEachCallback;
-import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.rules.MethodRule;
+import org.junit.runners.model.FrameworkMethod;
+import org.junit.runners.model.Statement;
 
 import io.mats3.MatsFactory;
-import io.mats3.test.jupiter.Extension_Mats;
+import io.mats3.test.abstractunit.AbstractMatsAnnotatedClass;
 
 /**
  * Helper class to test classes annotated with Mats annotations.
@@ -29,44 +29,45 @@ import io.mats3.test.jupiter.Extension_Mats;
  * This will still allow for an integration test of the Mats endpoint, checking annotations, and that state and
  * dtos can be serialized if needed.
  *
- * @author Ståle Undheim, 2023.09.13 stale.undheim@storebrand.no
+ * @author Ståle Undheim <stale.undheim@storebrand.no> 2024-11-21
  */
-public final class Extension_MatsSpring extends AbstractMatsSpring implements BeforeEachCallback, AfterEachCallback {
+public class Rule_MatsAnnotatedClass extends AbstractMatsAnnotatedClass implements MethodRule {
 
-
-    private Extension_MatsSpring(Extension_Mats extensionMats) {
-        super(extensionMats);
+    private Rule_MatsAnnotatedClass(Rule_Mats ruleMats) {
+        super(ruleMats);
     }
 
     /**
-     * Create a new Extension_MatsSpring instance, register to Junit using
-     * {@link org.junit.jupiter.api.extension.RegisterExtension}.
-     * @param extensionMats {@link Extension_Mats} to read the {@link MatsFactory} from.
-     * @return a new {@link Extension_MatsSpring}
+     * Create a new Rule_MatsSpring instance, register to Junit using
+     * {@link org.junit.Rule}.
+     * @param ruleMats {@link Rule_Mats} to read the {@link MatsFactory} from.
+     * @return a new {@link Rule_MatsAnnotatedClass}
      */
-    public static Extension_MatsSpring create(Extension_Mats extensionMats) {
-        return new Extension_MatsSpring(extensionMats);
+    public static Rule_MatsAnnotatedClass create(Rule_Mats ruleMats) {
+        return new Rule_MatsAnnotatedClass(ruleMats);
     }
 
     /**
      * Add classes to act as a source for annotations to register Mats endpoints for each test.
+     *
      */
-    @Override
-    public Extension_MatsSpring withClasses(Class<?>... annotatedMatsClasses) {
-        super.withClasses(annotatedMatsClasses);
+    public Rule_MatsAnnotatedClass withClasses(Class<?>... annotatedMatsClasses) {
+        super.addClasses(annotatedMatsClasses);
         return this;
     }
 
     @Override
-    @SuppressWarnings("PMD.AvoidAccessibilityAlteration")
-    // AvoidAccessibilityAlteration - need to reflect into test class to get fields.
-    public void beforeEach(ExtensionContext extensionContext) {
-        beforeEach(extensionContext.getTestInstance().orElse(null));
+    public Statement apply(Statement base, FrameworkMethod method, Object target) {
+        return new Statement() {
+            public void evaluate() throws Throwable {
+                beforeEach(target);
+                try {
+                    base.evaluate();
+                }
+                finally {
+                    afterEach();
+                }
+            }
+        };
     }
-
-    @Override
-    public void afterEach(ExtensionContext context) {
-        afterEach();
-    }
-
 }
